% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_data.R
\name{browse_rba_series}
\alias{browse_rba_series}
\alias{browse_rba_tables}
\title{Browse available RBA data series}
\usage{
browse_rba_series(search_string = "")

browse_rba_tables(search_string = "")
}
\arguments{
\item{search_string}{Word or phrase to search for, such as "gold" or "commodity" or "labour".
If left as \code{""}, the function will return all series.}
}
\value{
A \code{data.frame} (\code{tbl_df}) containing RBA data series/tables that match the \code{search_string}.
}
\description{
Browse available RBA data series
}
\examples{

# Find series that contain 'unemployment'
browse_rba_series("unemployment")

# Or all labour-related series
browse_rba_series("labour")

# Or those related to commodities
browse_rba_series("commodities")

# Or all series
browse_rba_series()

# Or just look for tables that contain the word 'labour'
browse_rba_tables("labour")

# Or all tables
browse_rba_tables()
}
