% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_rba_ts.R
\name{check_if_rba_ts}
\alias{check_if_rba_ts}
\title{Perform a minimal check to see if a dataframe is formatted in a standard
way for an RBA time series spreadsheet.
Checks to see if various strings are present in the first column of the DF.}
\usage{
check_if_rba_ts(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Logical; \code{TRUE} if \code{df} contains the expected strings in first col;
\code{FALSE} if not.
}
\description{
Perform a minimal check to see if a dataframe is formatted in a standard
way for an RBA time series spreadsheet.
Checks to see if various strings are present in the first column of the DF.
}
\keyword{internal}
