% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rba.R
\name{download_rba}
\alias{download_rba}
\title{Download statistical table(s) from the RBA}
\usage{
download_rba(urls, path = tempdir())
}
\arguments{
\item{urls}{Table url(s)}

\item{path}{Directory in which to save the downloaded file(s)}
}
\value{
Invisibly returns path to downloaded file(s)
}
\description{
Download statistical table(s) from the RBA
}
\examples{
\dontrun{
download_rba(url = "https://rba.gov.au/statistics/tables/xls/f02d.xls")
}

}
