% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rba_sheet.R
\name{excel_sheets_noguess}
\alias{excel_sheets_noguess}
\title{Wrapper around readxl::read_excel() that first checks to see that the
file format matches the file signature, and renames the file if not}
\usage{
excel_sheets_noguess(filename)
}
\arguments{
\item{filename}{Filename incl path}
}
\description{
Wrapper around readxl::read_excel() that first checks to see that the
file format matches the file signature, and renames the file if not
}
\keyword{internal}
