% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forecasts}
\alias{forecasts}
\title{Forecasts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25010 rows and 8 columns.
}
\usage{
forecasts
}
\description{
Forecasts
}
\keyword{datasets}
