% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rba_urls.R
\name{get_rba_urls}
\alias{get_rba_urls}
\title{Find the url(s) corresponding to RBA table number(s).
Note that multiple URLs can be returned for a single table_no, eg. \code{A3}.}
\usage{
get_rba_urls(table_no, cur_hist = "current")
}
\arguments{
\item{table_no}{Character vector of RBA table number(s), such as \code{"A1"}
or \code{c("A1", "g1")}.
Not case-sensitive.}

\item{cur_hist}{Either \code{"current"} (the default) or \code{"historical"}. Must be
length 1.
Choose between getting
the URLs corresponding to current RBA tables, or historical tables.}
}
\value{
Vector of URL(s) corresponding to the supplied \code{table_no}
}
\description{
Find the url(s) corresponding to RBA table number(s).
Note that multiple URLs can be returned for a single table_no, eg. \code{A3}.
}
\keyword{internal}
