% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba_forecasts.R
\name{latest_forecast_month}
\alias{latest_forecast_month}
\title{Obtain the month of the latest RBA SMP forecasts}
\usage{
latest_forecast_month()
}
\description{
Obtain the month of the latest RBA SMP forecasts
}
\details{
This function returns a length-one date, corresponding to first day of the
month of the latest RBA Statement on Monetary Policy forecasts.
}
\examples{
\dontrun{
latest_forecast_month()
}
}
\keyword{internal}
