% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rba.R
\name{load_rba}
\alias{load_rba}
\title{Load an Excel sheet containing an RBA statistical table}
\usage{
load_rba(filename)
}
\arguments{
\item{filename}{Filename, including path, of an RBA Excel spreadsheet}
}
\description{
Load an Excel sheet containing an RBA statistical table
}
