% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rba_sheet.R
\name{load_rba_sheet}
\alias{load_rba_sheet}
\title{Load the data sheet(s) from an RBA Excel file}
\usage{
load_rba_sheet(filename)
}
\arguments{
\item{filename}{Filename, including path, to an RBA Excel workbook}
}
\value{
list of dataframes
}
\description{
Load the data sheet(s) from an RBA Excel file
}
\details{
Function loads every sheet in the Excel workbook at \code{filename}
other than any with names that correspond to known non-data sheets
(eg. "Notes").
}
\keyword{internal}
