% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelim_tidy_rba.R
\name{prelim_tidy_old_f16}
\alias{prelim_tidy_old_f16}
\title{Function to wrangle historical yields data to get it in the standard format
Called indirectly from tidy_rba()}
\usage{
prelim_tidy_old_f16(excel_sheet)
}
\arguments{
\item{excel_sheet}{Excel sheet with no tidying done}
}
\description{
Function to wrangle historical yields data to get it in the standard format
Called indirectly from tidy_rba()
}
\keyword{internal}
