% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba_forecasts.R
\name{rba_forecasts}
\alias{rba_forecasts}
\title{Compile the RBA's public forecasts of key economic variables over time}
\usage{
rba_forecasts(refresh = TRUE)
}
\arguments{
\item{refresh}{logical; default is \code{TRUE}. When set to \code{TRUE}, the RBA's
website will be scraped to obtain recent forecasts. When \code{FALSE}, only
the package's internal data will be returned, which may be out of date.}
}
\description{
Compile the RBA's public forecasts of key economic variables over time
}
