% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba_value_to_num.R
\name{rba_value_to_num}
\alias{rba_value_to_num}
\title{Convert RBA values to numeric}
\usage{
rba_value_to_num(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Some RBA data are given as ranges (eg. 3-4\%). We want to take the
mean of the range (3.5\%). This primarily applies to forecasts, but also
includes some actual data (such as the cash rate target in the early 1990s).
}
\details{
Other values have non-standard characters such as
em-dashes rather than minus signs. We want to replace these and convert
to numeric.
}
\examples{
rba_value_to_num(c("1.0", "-2", "1½", "1½–2½", "1½–2½", "1½–2½", "1½–2½"))
}
\keyword{internal}
