% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cashrate.R
\name{read_cashrate_target}
\alias{read_cashrate_target}
\title{Convenience function to get the RBA's cash rate target}
\usage{
read_cashrate_target(dates = c("all", "changes"))
}
\arguments{
\item{dates}{Either \code{all} or \code{changes}. If \code{all} (the default),
returns a data frame with rows for each date between 23 January 1990
and \code{Sys.Date()}. If \code{changes}, returns a data frame with rows only for
those dates on which the cash rate changed.}
}
\value{
A \code{tbl_df} with two columns: \code{date} and \code{cash_rate}.
}
\description{
Convenience function to get the RBA's cash rate target
}
\keyword{internal}
