% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rba_sheet.R
\name{read_excel_noguess}
\alias{read_excel_noguess}
\title{Drop-in replacement for \code{readxl::read_excel()} that does not infer
file type based on extension}
\usage{
read_excel_noguess(path, ...)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{...}{Arguments passed to \code{readxl::read_xls()} or \code{readxl::read_xlsx()}}
}
\description{
Drop-in replacement for \code{readxl::read_excel()} that does not infer
file type based on extension
}
\keyword{internal}
