% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rba.R
\name{read_rba}
\alias{read_rba}
\title{Download, import, and tidy statistical tables from the RBA}
\usage{
read_rba(table_filenames, path = tempdir())
}
\arguments{
\item{table_filenames}{character vector of table filename(s) without extension,
such as "g01hist"}

\item{path}{directory in which to save file(s); default is \code{tempdir()}}
}
\value{
A single tidy tibble containing the requested table(s)
}
\description{
Download, import, and tidy statistical tables from the RBA
}
\examples{
\dontrun{
read_rba(table_filenames = c("g01hist", "g03hist"))
}
}
