% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rba.R
\name{read_rba_local}
\alias{read_rba_local}
\title{Load and tidy local RBA Excel sheets}
\usage{
read_rba_local(filenames)
}
\arguments{
\item{filenames}{Vector of filenames (with path) pointing to local RBA Excel sheets}
}
\description{
Load and tidy local RBA Excel sheets
}
\examples{
\dontrun{
read_rba_local("data/rba_file.xls")
}
}
