% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readrba-package.R
\docType{package}
\name{readrba-package}
\alias{readrba}
\alias{readrba-package}
\title{readrba: Download and Tidy Statistical Tables from the Reserve Bank of Australia}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Download data from the Reserve Bank of Australia in tidy
  tibble format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MattCowgill/readrba}
  \item Report bugs at \url{https://github.com/MattCowgill/readrba/issues}
}

}
\author{
\strong{Maintainer}: Matt Cowgill \email{mattcowgill@gmail.com} (\href{https://orcid.org/0000-0003-0422-3300}{ORCID})

}
\keyword{internal}
