% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safely_read_html.R
\name{safely_read_html}
\alias{safely_read_html}
\title{\code{read_html()} safely.}
\usage{
safely_read_html(url, ...)
}
\arguments{
\item{url}{url to read with \code{xml2::read_html()}}

\item{...}{arguments passed to \code{xml2::read_html()}}
}
\value{
If the URL is read without error, an XML document
(see \code{?xml2::read_html}).
}
\description{
The purpose of this function is to specify a user agent for scraping,
to try a second time if a scrape attempt fails,
and to return a useful error if the page cannot be scraped.
}
\keyword{internal}
