% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba_forecasts.R
\name{scrape_rba_forecasts}
\alias{scrape_rba_forecasts}
\title{Import and tidy forecasts from the published RBA SMP .xlsx file
Not intended to be called directly; call from \code{read_forecasts()}}
\usage{
scrape_rba_forecasts()
}
\description{
Import and tidy forecasts from the published RBA SMP .xlsx file
Not intended to be called directly; call from \code{read_forecasts()}
}
\keyword{internal}
