% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_table_list.R
\name{scrape_table_list}
\alias{scrape_table_list}
\title{Scrape the RBA site to obtain links to tables}
\usage{
scrape_table_list()
}
\value{
A tibble containing the text and URL of XLS/XLSX links
}
\description{
Scrape the RBA site to obtain links to tables
}
