% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rba.R
\name{tables_from_seriesid}
\alias{tables_from_seriesid}
\title{Given series ID(s), find the corresponding table number(s)}
\usage{
tables_from_seriesid(series_id)
}
\arguments{
\item{series_id}{A character vector of RBA series ID(s)}
}
\description{
Given series ID(s), find the corresponding table number(s)
}
\keyword{internal}
