% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rba.R
\name{tidy_rba_normal}
\alias{tidy_rba_normal}
\title{Function to tidy RBA sheets that are formatted in the standard way}
\usage{
tidy_rba_normal(excel_sheet, .table_title)
}
\arguments{
\item{excel_sheet}{Data.frame of an RBA spreadsheet}

\item{.table_title}{Length 1 character vector of table title}
}
\description{
Function to tidy RBA sheets that are formatted in the standard way
}
\keyword{internal}
