% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_exists.R
\name{url_exists}
\alias{url_exists}
\alias{check_url_success}
\title{Check if URLs are valid}
\usage{
url_exists(urls)

check_url_success(url)
}
\arguments{
\item{urls}{Character vector containing url(s)}

\item{url}{Character vector of length 1; a url incl. http/https, such as
\verb{https://www.google.com}.}
}
\value{
Logical vector; names of the vector are the supplied urls

Logical. If status category = "Success", \code{TRUE}, else \code{FALSE}.
}
\description{
Check if URLs are valid

Check a single URL's status
Internal function used by \code{url_exists()}
}
\keyword{internal}
