% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rba.R
\name{download_rba}
\alias{download_rba}
\title{Download statistical table(s) from the RBA}
\usage{
download_rba(table_filenames, path = tempdir())
}
\arguments{
\item{table_filenames}{Table filename(s) without extension (eg. "d01hist")}

\item{path}{Directory in which to save the downloaded file(s)}
}
\value{
Invisibly returns path to downloaded file(s)
}
\description{
Download statistical table(s) from the RBA
}
\examples{
\dontrun{
download_rba(table_filenames = "d01hist")
}

}
