% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rba.R
\name{download_rba}
\alias{download_rba}
\alias{do_download_files}
\title{Download statistical table(s) from the RBA}
\usage{
download_rba(urls, path = tempdir())

do_download_files(urls, filenames_with_path)
}
\arguments{
\item{urls}{Table url(s)}

\item{path}{Directory in which to save the downloaded file(s);
default is \code{tempdir()}}
}
\value{
Invisibly returns path to downloaded file(s)
}
\description{
Download statistical table(s) from the RBA

Internal function to download files
}
\examples{
\dontrun{
download_rba(urls = "https://rba.gov.au/statistics/tables/xls/f02d.xls")
}

}
\keyword{internal}
