% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rba.R
\name{read_rba}
\alias{read_rba}
\title{Download, import, and tidy statistical tables from the RBA}
\usage{
read_rba(table_no = NULL, cur_hist = "current", path = tempdir())
}
\arguments{
\item{table_no}{Table number(s) as character vector, such as "A1" or c("a1.1", "g1").}

\item{cur_hist}{Character; either "current" or "historical".}

\item{path}{directory in which to save file(s); default is \code{tempdir()}}
}
\value{
A single tidy tibble containing the requested table(s)
}
\description{
Download, import, and tidy statistical tables from the RBA
}
\examples{
\dontrun{
read_rba(table_no = c("a1.1", "g1"))
}
}
