% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readrba-package.R
\docType{package}
\name{readrba-package}
\alias{readrba}
\alias{readrba-package}
\title{readrba: Download and Tidy Data from the Reserve Bank of Australia}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Download up-to-date data from the Reserve Bank of Australia 
    in a tidy data frame. Package includes functions to download current and 
    historical RBA statistical tables 
    (<https://www.rba.gov.au/statistics/tables/>) and RBA forecasts 
    (<https://www.rba.gov.au/publications/smp/forecasts-archive.html>). Data
    includes a broad range of Australian macroeconomic and financial time
    series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mattcowgill.github.io/readrba/index.html}
  \item Report bugs at \url{https://github.com/MattCowgill/readrba/issues}
}

}
\author{
\strong{Maintainer}: Matt Cowgill \email{mattcowgill@gmail.com} (\href{https://orcid.org/0000-0003-0422-3300}{ORCID})

}
\keyword{internal}
