% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rba_sheet.R
\name{rename_excel}
\alias{rename_excel}
\title{This function checks to see if an Excel file's extension matches its
signature; if it does not, the file will be renamed so that the file
extension is equal to the signature}
\usage{
rename_excel(filename)
}
\description{
This function checks to see if an Excel file's extension matches its
signature; if it does not, the file will be renamed so that the file
extension is equal to the signature
}
\keyword{internal}
