% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rba_forecasts.R
\name{scrape_rba_forecasts}
\alias{scrape_rba_forecasts}
\title{Scrape the RBA's website to obtain recent forecasts}
\usage{
scrape_rba_forecasts()
}
\value{
A tidy \code{tbl_df} containing 7 columns:
\itemize{
\item{\code{forecast_date}}{ The (approximate) date on which the forecast was published. Note that this is the first day of the publication month, so the \code{forecast_date} for forecasts in the February 2020 Statement on Monetary Policy is \code{as.Date("2020-02-01")}.}
\item{\code{date}}{ The date to which the forecast pertains. Note that this is the first day of the final month of the relevant quarter. For example, a forecast of GDP in the June quarter 2021 will be \code{as.Date("2021-06-01")}.}
\item{\code{series}}{ Short, snake_case description of the data series being forecast, such as \code{gdp_change} or \code{unemp_rate}.}
\item{\code{value}}{ The forecast value, in per cent. For example, if GDP growth is forecast to be 3 per cent, the value will be \code{3}.}
\item{\code{series_desc}}{ Full description of the series being forecast, as per the RBA website, such as "Real household disposable income".}
\item{\code{source}}{ For recent forecasts, this is 'SMP', meaning the RBA's Statement on Monetary Policy.}
\item{\code{notes}}{ Notes accompanying the forecasts, as per the RBA's website. Note these are identical for item in a given \code{forecast_date}.}
}
}
\description{
Scrape and tidy forecasts from all editions of the RBA's Statement on Monetary Policy
from November 2018 onwards.
}
\keyword{internal}
