% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rba.R
\name{tidy_rba}
\alias{tidy_rba}
\title{Tidy a statistical table from the RBA}
\usage{
tidy_rba(excel_sheet, series_id = NULL)
}
\arguments{
\item{excel_sheet}{Dataframe of RBA spreadsheet.}

\item{series_id}{Optional series ID}
}
\value{
Tidy tibble
}
\description{
Tidy a statistical table from the RBA
}
